refactor(cp): rename directories
