+++ src/core/copy/prompt.ts
@@ -1,4 +1,4 @@
-import { Directory } from "@/utils/projects";
+import { Directory } from "@/utils/directory";
 import { select } from "@inquirer/prompts";
 
 /**
