+++ test/commands.test.ts
@@ -1,11 +1,8 @@
 import { expect, spyOn, test } from "bun:test";
 
 import * as commander from "commander";
-import {
-	commandOptions,
-	commanderInit,
-	configureCommand,
-} from "../src/utils/commands";
+import { configureCommand } from "../src/utils/command";
+import { commandOptions, commanderInit } from "@/core/copy/command";
 
 test("should configure Commander program with specified command options", () => {
 	const program = new commander.Command();
